refactor(models): consolidate Timing family file layout per spec packages (Rule 0007 A)
