+++ tests/test_armodel/models/M2/AUTOSARTemplates/CommonStructure/Timing/TimingConstraint/test_SynchronizationTiming.py
@@ -4,7 +4,7 @@ AUTOSAR CommonStructure.Timing.TimingConstraint module.
 """
 
 from armodel.models.M2.AUTOSARTemplates.AutosarTopLevelStructure import AUTOSAR
-from armodel.models.M2.AUTOSARTemplates.CommonStructure.Timing.TimingConstraint.SynchronizationTimingConstraint import (
+from armodel.models.M2.AUTOSARTemplates.CommonStructure.Timing.TimingConstraint.SynchronizationTiming import (
     EventOccurrenceKindEnum,
     SynchronizationTimingConstraint,
     SynchronizationTypeEnum,
